Fix 'whenCalledDoes' for IO and output parameters
This approach works for both IO and output parameters, as the mocked method's signature handles them correctly.
The scope of the issue was limited to our call into 'executeTransientMethod,' where we didn’t define 'params' as IO, which caused the output parameter not to be updated. Now IO and output parameters in a mocked method work as intended.